+++ packages/nimble-components/src/theme-provider/tests/theme-provider.spec.ts
@@ -1,6 +1,6 @@
 import { spinalCase } from '@microsoft/fast-web-utilities';
 import * as designTokensNamespace from '../design-tokens';
-import { tokenNames } from '../design-token-names';
+import { tokenNames, suffixFromTokenName } from '../design-token-names';
 import { getSpecTypeByNamedList } from '../../utilities/tests/parameterized';
 import { ThemeProvider } from '..';
 
@@ -62,23 +62,6 @@ describe('Theme Provider', () => {
     });
 
     describe('design token has approved suffix', () => {
-        // Order of suffixes in the array matter, as we want single word suffixes after the multi-word ones
-        const tokenSuffixes = [
-            'RgbPartialColor',
-            'FontColor',
-            'FontLineHeight',
-            'FontWeight',
-            'FontSize',
-            'TextTransform',
-            'FontFamily',
-            'Font',
-            'Size',
-            'Width',
-            'Height',
-            'Delay',
-            'Padding',
-            'Color'
-        ];
         const propertyNames = designTokenPropertyNames.map(
             (name: DesignTokenPropertyName) => ({ name })
         );
@@ -92,8 +75,8 @@ describe('Theme Provider', () => {
             );
             specType(`for token name ${propertyName.name}`, () => {
                 expect(
-                    tokenSuffixes.some(tokenSuffix => propertyName.name.endsWith(tokenSuffix))
-                ).toBeTrue();
+                    suffixFromTokenName(propertyName.name)
+                ).not.toBeUndefined();
             });
         }
     });